+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next
 - Add drag-and-drop, selection and context menus in queue panel
 - Add warning when adding duplicate songs to a playlist
+- Scroll sidebar to opened playlist
 - Improve queue panel performance
 - Don't clear selection unnecessarily
 - Reload file when updating the currently playing song's info